rearrange tests to add file specific to definemap and definelist
